Introduce Router
- Introduce React Router
- Move Main Content to its own Component
- Break out some constants to a shared file
