+++ package.json
@@ -17,6 +17,7 @@
     "react-prism": "^4.3.0",
     "react-redux": "^5.0.6",
     "react-responsive-picture": "^2.0.0",
+    "react-router-dom": "^5.0.1",
     "react-scroll": "^1.5.4",
     "redux": "^3.7.2",
     "redux-devtools-extension": "^2.13.2",
