+++ src/shared.js
@@ -1,0 +1,23 @@
+const sections = {
+    HOME: 'home',
+    SLACK: 'slack',
+    GITHUB: 'github',
+    MAILING_LIST: 'mailing-list',
+    LEARN: 'learn',
+    USES: 'uses',
+    DESCRIPTION: 'description',
+    TEAM: 'team',
+    DEFINITION: 'definition',
+    ADVISORS: 'advisors',
+    PARTNERS: 'partners',
+    SUPPORTERS: 'supporters',
+    HIRING: 'hiring'
+}
+
+const WHITEPAPER_URL = 'https://keep.network/whitepaper'
+
+
+export {
+    sections,
+    WHITEPAPER_URL
+}
